Avoid lib/services/local/events.go exploding in size
