+++ lib/services/local/workload_identity_x509_revocation.go
@@ -18,9 +18,11 @@ package local
 
 import (
 	"context"
+	"strings"
 
 	"github.com/gravitational/trace"
 
+	apidefaults "github.com/gravitational/teleport/api/defaults"
 	workloadidentityv1pb "github.com/gravitational/teleport/api/gen/proto/go/teleport/workloadidentity/v1"
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/lib/backend"
@@ -123,4 +125,44 @@ func (b *WorkloadIdentityX509RevocationService) UpdateWorkloadIdentityX509Revoca
 ) (*workloadidentityv1pb.WorkloadIdentityX509Revocation, error) {
 	updated, err := b.service.ConditionalUpdateResource(ctx, resource)
 	return updated, trace.Wrap(err)
+}
+
+func newWorkloadIdentityX509RevocationParser() *workloadIdentityX509RevocationParser {
+	return &workloadIdentityX509RevocationParser{
+		baseParser: newBaseParser(backend.NewKey(workloadIdentityX509RevocationPrefix)),
+	}
+}
+
+type workloadIdentityX509RevocationParser struct {
+	baseParser
+}
+
+func (p *workloadIdentityX509RevocationParser) parse(event backend.Event) (types.Resource, error) {
+	switch event.Type {
+	case types.OpDelete:
+		name := event.Item.Key.TrimPrefix(backend.NewKey(workloadIdentityX509RevocationPrefix)).String()
+		if name == "" {
+			return nil, trace.NotFound("failed parsing %v", event.Item.Key.String())
+		}
+
+		return &types.ResourceHeader{
+			Kind:    types.KindWorkloadIdentityX509Revocation,
+			Version: types.V1,
+			Metadata: types.Metadata{
+				Name:      strings.TrimPrefix(name, backend.SeparatorString),
+				Namespace: apidefaults.Namespace,
+			},
+		}, nil
+	case types.OpPut:
+		resource, err := services.UnmarshalWorkloadIdentityX509Revocation(
+			event.Item.Value,
+			services.WithExpires(event.Item.Expires),
+			services.WithRevision(event.Item.Revision))
+		if err != nil {
+			return nil, trace.Wrap(err, "unmarshalling resource from event")
+		}
+		return types.Resource153ToLegacy(resource), nil
+	default:
+		return nil, trace.BadParameter("event %v is not supported", event.Type)
+	}
 }
